requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
pymupdf>=1.24.0
paddleocr>=2.7.3
paddlepaddle>=2.6.2
Pillow>=10.4.0
pydantic>=2.0.0
numpy
